# [ENHANCEMENT] TSDB: Check CRC without allocating