Use the existing utility function which does this.

Picking one benchmark at random:
```
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/promql
cpu: Intel(R) Core(TM) i7-14700K
                                      │ before.txt  │          after.txt           │
                                      │   sec/op    │   sec/op     vs base         │
RangeQuery/expr=a_hundred,steps=100-4   441.1µ ± 2%   438.9µ ± 3%  ~ (p=0.937 n=6)

                                      │  before.txt  │             after.txt              │
                                      │     B/op     │     B/op      vs base              │
RangeQuery/expr=a_hundred,steps=100-4   84.52Ki ± 0%   82.21Ki ± 0%  -2.74% (p=0.002 n=6)

                                      │ before.txt  │             after.txt              │
                                      │  allocs/op  │  allocs/op   vs base               │
RangeQuery/expr=a_hundred,steps=100-4   1.300k ± 0%   1.104k ± 0%  -15.08% (p=0.002 n=6)
```