@machine424 2024-03-11
good catch! You could have waited until June 1 https://github.com/prometheus/prometheus/pull/10789 :)
lgtm.