@bboreham 2024-03-11
Ha!  I forgot about that; previous was `Reader.Chunk()` and this is `ChunkDiskMapper.Chunk()`.